Make install_exercise_icons idempotent, use Khan repo & other cleanup
Reviewers: csilvers

Reviewed By: csilvers

Differential Revision: http://phabricator.khanacademy.org/D7873

+++ internal-webserver/setup.sh
@@ -174,7 +174,7 @@ install_repo_backup() {
     ssh -oStrictHostKeyChecking=no khanacademy.kilnhg.com >/dev/null 2>&1 || true
     echo "Visit https://khanacademy.kilnhg.com/Keys"
     echo "Log in as user ReadOnlyKiln (ask kamens for the password),"
-    echo "   click 'Add a New Key', paste the contents of ~/.ssh/id_rsa.pub"
+    echo "   click 'Add a New Key', paste the contents of $HOME/.ssh/id_rsa.pub"
     echo "   into the box, and hit 'Save key'"
 }
 
@@ -311,7 +311,7 @@ install_gae_default_version_notifier() {
     echo "Installing gae-default-version-notifier"
     git clone git://github.com/Khan/gae-default-version-notifier.git || \
         ( cd gae-default-version-notifier && git pull )
-    echo "For now, set up ~/gae-default-version-notifier/secrets.py based"
+    echo "For now, set up $HOME/gae-default-version-notifier/secrets.py based"
     echo "on secrets.py.example and the 'real' secrets.py."
     # TODO(csilvers): instead, control this via monit(1).
     echo "Then run: nohup python notify.py </dev/null >/dev/null 2>&1 &"
@@ -336,7 +336,7 @@ install_publish_notifier() {
     echo "Installing publish-notifier"
     git clone git://github.com/Khan/publish-notifier.git || \
         ( cd publish-notifier && git pull )
-    echo "For now, set up ~/publish-notifier/secrets.py based"
+    echo "For now, set up $HOME/publish-notifier/secrets.py based"
     echo "on secrets.py.example and the 'real' secrets.py."
     # TODO(csilvers): instead, control this via monit(1).
     echo "Then run: nohup python notify.py </dev/null >/dev/null 2>&1 &"
@@ -351,38 +351,55 @@ install_kahntube_ouath_collector() {
     sudo update-rc.d -f khantube-oauth-collector-daemon remove
     sudo ln -snf "${HOME}"/aws-config/internal-webserver/etc/init.d/khantube-oauth-collector-daemon /etc/init.d
     sudo update-rc.d khantube-oauth-collector-daemon defaults
-    echo "Add ~/internal-webserver/khantube-oauth-collector/secrets.py "
-    echo "if it is not already there, according to the secrets_example.py in "
-    echo "the same directory."
-    echo "Hit <enter> when this is done:"
-    read prompt
+    if [ ! -e "$HOME/internal-webserver/khantube-oauth-collector/secrets.py" ]; then
+        echo "Add $HOME/internal-webserver/khantube-oauth-collector/secrets.py "
+        echo "according to the secrets_example.py in the same directory."
+        echo "Hit <enter> when this is done:"
+        read prompt
+    fi
     sudo service khantube-oauth-collector-daemon restart
 }
 
 install_exercise_icons() {
     # A utility to generate exercise icons. Currently used at http://khanacademy.org/commoncore/map.
-    # https://github.com/jnetterf/exercise-icons/
-    sudo aptitude install gcc-multilib xdg-utils libxml2-dev libcurl4-openssl-dev imagemagick
-    wget http://downloads.dlang.org/releases/2014/dmd_2.065.0-0_amd64.deb -O /tmp/dmd.deb
-    sudo dpkg -i /tmp/dmd.deb
-    
-    cd /usr/local/share
-    sudo wget https://phantomjs.googlecode.com/files/phantomjs-1.9.0-linux-x86_64.tar.bz2
-    sudo tar -xjf /usr/local/share/phantomjs-1.9.0-linux-x86_64.tar.bz2
-    sudo ln -sf /usr/local/share/phantomjs-1.9.0-linux-x86_64/bin/phantomjs /usr/local/bin/phantomjs
-    sudo git clone git://github.com/n1k0/casperjs.git /usr/local/src/casperjs
-    cd /usr/local/src/casperjs
-    sudo git checkout tags/1.0.2
-    sudo ln -sf /usr/local/src/casperjs/bin/casperjs /usr/local/bin/casperjs
-    
-    git clone git@github.com:jnetterf/exercise-icons.git
-    echo "Add ~/exercise-icons/secrets.txt"
-    echo "if it is not already there, according to the instructions in README.md."
-    echo "BUCKET should be set to 'ka-exercise-screenshots-2'."
-    echo "Hit <enter> when this is done:"
-    read prompt
-    cd exercise-icons
-    make
+    # https://github.com/Khan/exercise-icons/
+    sudo aptitude -y install gcc-multilib xdg-utils libxml2-dev libcurl4-openssl-dev imagemagick
+
+    if [ ! -e "/usr/bin/dmd" ]; then
+        wget http://downloads.dlang.org/releases/2014/dmd_2.065.0-0_amd64.deb -O /tmp/dmd.deb
+        sudo dpkg -i /tmp/dmd.deb
+        rm /tmp/dmd.deb
+    fi
+
+    (
+        cd /usr/local/share
+
+        if [ ! -L "/usr/local/bin/phantomjs" ]; then
+		sudo wget https://phantomjs.googlecode.com/files/phantomjs-1.9.0-linux-x86_64.tar.bz2
+		sudo tar -xjf /usr/local/share/phantomjs-1.9.0-linux-x86_64.tar.bz2
+		sudo rm /usr/local/share/phantomjs-1.9.0-linux-x86_64.tar.bz2
+		sudo ln -sf /usr/local/share/phantomjs-1.9.0-linux-x86_64/bin/phantomjs /usr/local/bin/phantomjs
+        fi
+        if [ ! -L "/usr/local/bin/casperjs" ]; then
+		sudo git clone git://github.com/n1k0/casperjs.git /usr/local/src/casperjs
+		cd /usr/local/src/casperjs/
+		sudo git fetch origin
+		sudo git checkout tags/1.0.2
+		sudo ln -snf /usr/local/src/casperjs/bin/casperjs /usr/local/bin/casperjs
+        fi
+
+        cd "$HOME"
+        git clone git@github.com:Khan/exercise-icons.git || (cd exercise-icons && git pull)
+        if [ ! -e "$HOME/exercise-icons/secrets.txt" ]; then
+            echo "Add $HOME/exercise-icons/secrets.txt"
+            echo "according to the instructions in README.md."
+            echo "BUCKET should be set to 'ka-exercise-screenshots-2'."
+            echo "Hit <enter> when this is done:"
+            read prompt
+        fi
+        cd exercise-icons
+        make
+    )
 }
 
 cd "$HOME"
@@ -393,7 +410,7 @@ install_root_config_files
 install_user_config_files
 install_appengine
 install_repo_backup
-##install_gerrit
+#install_gerrit
 install_phabricator
 install_gae_default_version_notifier
 install_beep_boop
